Add a question of the day panel
Add a question of the day panel to the right hand side with the other widgets. This gets random questions with a score >7 and older than a year by tags. The top 3 tags are set to `code-golf`, `KOTH`, and `fastest-code`. The other 2 tags are rotated daily and are from the first page of tags by popularity.