Fix for some older headers
The language detection for headers such as "Python (80)" stopped working at some point, so this is a fix for that. Note that in the ambiguous case of "Language 6" (could either be "Language 6" or a 6-byte "Language" entry), the number is taken as part of the language name if there is only one digit.